pyproject.toml
[project]
name = "retina"
version = "0.1.0"
description = "Scripts used in RETINA facility @ PoliMi"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "astra-toolbox>=2.4.0",
    "astropy>=7.1.0",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "scipy>=1.16.2",
    "tifffile>=2025.9.20",
    "torch>=2.0.1",
    "tqdm>=4.67.1",
]
